refactor: reorg models folder
