+++ cognite/neat/rules/models/_base.py
