+++ cognite/neat/rules/models/_information_rules.py
