+++ cognite/neat/rules/models/_types/_base.py
