+++ cognite/neat/rules/models/dms/__init__.py
@@ -1,0 +1,4 @@
+from ._dms_architect_rules import DMSRules
+from ._dms_schema import DMSSchema
+
+__all__ = ["DMSRules", "DMSSchema"]
